Acknowledge reception of data in `TrinoResult`
A query only transitions to a FINISHED state when the results are fully consumed. The reception of the data is acknowledged by calling the next_uri before exposing the data through dbapi.

`dbapi.execute()` will now block until the first result is received. This will ensure the result set is exhausted for DML queries and as such remove the need to for calling `dbapi.fetchall()`.