Do not fail on empty values for mutli-valued headers
Before this change user would receive a ValueError: not enough values to
unpack if the session property, role or prepared statements headers
existed but had empty values. While Trino itself doesn't send empty
headers it's possible for them to be present because of some proxy or
gateway sitting between client and the server.