Add support for trailing semicolon in sql statements
Naive implementation of removal of optional trailing semicolon.
This allows sql statements passed to `cursor.execute()` to contain an optional trailing semicolon.

preserves whitespace at the end of the statement.

Will not strip the trailing semicolon if there are trailing comments (edge case) - can be made more sophisticated if need be.