Refactor setup.py to avoid cyclic dependency on trino.__version__
__version__ is defined in __init__.py in trino module which means that
we cannot easily use the module version in the library code because it
introduces a cyclic dependency.

This change introduces a _version.py which also includes other
meta-information and changes setup.py to read those values from there
instead of inline.

The release script is also updated since the version now needs to be
updated in a different file.