Suppress failure when reading $partitions system table in get_indexes
Not all connectors have a `$partitions` table. This caused `get_indexes`
to fail when called on a non-Hive (or non-partitioned Hive) table.

Since Trino engine doesn't have concept of partitions there's no single
way to fetch partition columns. One option is to parse the output of
`SHOW CREATE TABLE` to identify them but the logic would differ based on
what connector is being used. So we just opt to suppress the failure in
case of a non-Hive or non-partitioned Hive table instead.